+++ src/schemas/users/users.service.ts
@@ -8,7 +8,9 @@ import { deleteFile, deleteFiles } from 'src/utils/storageProcess/deleteFiles';
 import { compare, hash } from 'bcrypt';
 import { filterNullsObject } from 'src/utils/helpers/filterNulls';
 import { JwtService } from '@nestjs/jwt';
-import { TokenPayload } from 'src/types/token-payload.type';
+import { FullTokenPayload, TokenPayload } from 'src/types/token-payload.type';
+import { Request } from 'express';
+import { UserRole } from 'src/enums/user-role.enum';
 
 @Injectable()
 export class UsersService {
@@ -18,10 +20,22 @@ export class UsersService {
     private jwtService: JwtService,
   ) {}
 
-  passwordRemover(user: User) {
+  private passwordRemover(user: User) {
     // eslint-disable-next-line @typescript-eslint/no-unused-vars
     const { password, ...rest } = user;
     return rest;
+  }
+
+  getUserTokenData(req: Request) {
+    const userTokenData = {
+      ...req?.user,
+    } as FullTokenPayload;
+
+    userTokenData['expiredIn'] = `${Math.floor(
+      (userTokenData.exp - userTokenData.iat) / 3600,
+    )} Hours`;
+
+    return userTokenData;
   }
 
   async getUsers(conditions: Record<string, any>, withPass: boolean = false) {
@@ -67,8 +81,23 @@ export class UsersService {
     }
   }
 
-  async createUser(createUserDto: CreateUserDto) {
+  async createUser(
+    createUserDto: CreateUserDto,
+    userTokenData: FullTokenPayload,
+  ) {
     try {
+      if (
+        createUserDto.role === UserRole.ADMIN &&
+        (!userTokenData || userTokenData.role === UserRole.MEMBER)
+      ) {
+        return {
+          message:
+            'Unauthorized entrance, you must be an admin to create another admin account',
+          data: { token: userTokenData },
+          status: 401,
+        };
+      }
+
       const hashedPass = await hash(createUserDto.password, 12);
 
       const newUser = this.userRepository.create({
@@ -92,7 +121,11 @@ export class UsersService {
     }
   }
 
-  async updateUser(id: string, updateUserDto: UpdateUserDto) {
+  async updateUser(
+    id: string,
+    updateUserDto: UpdateUserDto,
+    userTokenData: FullTokenPayload,
+  ) {
     try {
       const user = await this.getUserById(id);
       if (!user) {
@@ -100,6 +133,18 @@ export class UsersService {
           message: 'Invalid data',
           data: `Provided user does not exist`,
           status: 404,
+        };
+      }
+
+      if (
+        userTokenData.userId !== id &&
+        userTokenData.role !== UserRole.ADMIN
+      ) {
+        return {
+          message:
+            "Unauthorized entrance, you're only allowed to update your account",
+          data: { token: userTokenData, id },
+          status: 401,
         };
       }
 
@@ -148,7 +193,7 @@ export class UsersService {
     }
   }
 
-  async deleteUser(id: string) {
+  async deleteUser(id: string, userTokenData: FullTokenPayload) {
     try {
       const user = await this.getUserById(id);
       if (user.status === 404) {
@@ -159,10 +204,23 @@ export class UsersService {
         };
       }
 
+      if (
+        userTokenData.userId !== id &&
+        userTokenData.role !== UserRole.ADMIN
+      ) {
+        return {
+          message:
+            "Unauthorized entrance, you're only allowed to delete your account",
+          data: { token: userTokenData, id },
+          status: 401,
+        };
+      }
+
       const response = await this.userRepository.delete(id);
 
       // delete the image related to the file
-      deleteFile('./public/assets/users/' + user?.data?.avatar);
+      user?.data?.avatar &&
+        deleteFile('./public/assets/users/' + user?.data?.avatar);
 
       return {
         message: 'User has been deleted successfully',
@@ -185,7 +243,7 @@ export class UsersService {
         };
       }
 
-      const user = response.data[0];
+      const user: User = response.data[0];
       if (!(await compare(password, user?.password))) {
         return {
           message: 'Invalid password',
@@ -197,6 +255,8 @@ export class UsersService {
       const payload: TokenPayload = {
         userId: user?.id,
         username: user?.userName,
+        role: user?.role,
+        email: user?.email,
       };
 
       return {
